Address comments and clean up
